code refactoring: Folder Entity 폴더 구조 분리
