+++ src/main/java/com/aisip/OnO/backend/problem/repository/ProblemRepository.java
@@ -1,4 +1,4 @@
-package com.aisip.OnO.backend.problem.repository.problem;
+package com.aisip.OnO.backend.problem.repository;
 
 import com.aisip.OnO.backend.problem.entity.Problem;
 import org.springframework.data.jpa.repository.JpaRepository;
