+++ src/main/java/com/aisip/OnO/backend/problem/service/ProblemService.java
@@ -5,15 +5,15 @@ import com.aisip.OnO.backend.fileupload.service.FileUploadService;
 import com.aisip.OnO.backend.problem.dto.ProblemDeleteRequestDto;
 import com.aisip.OnO.backend.problem.dto.ProblemImageDataRegisterDto;
 import com.aisip.OnO.backend.problem.dto.ProblemRegisterDto;
-import com.aisip.OnO.backend.problem.entity.Folder;
+import com.aisip.OnO.backend.folder.entity.Folder;
 import com.aisip.OnO.backend.problem.entity.ProblemImageData;
-import com.aisip.OnO.backend.problem.exception.FolderErrorCase;
+import com.aisip.OnO.backend.folder.exception.FolderErrorCase;
 import com.aisip.OnO.backend.problem.exception.ProblemErrorCase;
-import com.aisip.OnO.backend.problem.repository.folder.FolderRepository;
-import com.aisip.OnO.backend.problem.repository.problem.ProblemImageDataRepository;
+import com.aisip.OnO.backend.folder.repository.FolderRepository;
+import com.aisip.OnO.backend.problem.repository.ProblemImageDataRepository;
 import com.aisip.OnO.backend.problem.dto.ProblemResponseDto;
 import com.aisip.OnO.backend.problem.entity.Problem;
-import com.aisip.OnO.backend.problem.repository.problem.ProblemRepository;
+import com.aisip.OnO.backend.problem.repository.ProblemRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
